+++ src/treelearner/serial_tree_learner.h
@@ -293,10 +293,10 @@ class SerialTreeLearner: public TreeLearner {
   /*! \brief map dataset to leaves, used for linear model */
   std::vector<int> leaf_map_;
   /*! \brief temporary storage for calculating linear model */
-  std::vector<std::vector<double>> XTHX_;
-  std::vector<std::vector<double>> XTg_;
-  std::vector<std::vector<std::vector<double>>> XTHX_by_thread_;
-  std::vector<std::vector<std::vector<double>>> XTg_by_thread_;
+  std::vector<std::vector<float>> XTHX_;
+  std::vector<std::vector<float>> XTg_;
+  std::vector<std::vector<std::vector<float>>> XTHX_by_thread_;
+  std::vector<std::vector<std::vector<float>>> XTg_by_thread_;
 };
 
 inline data_size_t SerialTreeLearner::GetGlobalDataCountInLeaf(int leaf_idx) const {
